+++ packages/core/src/baseFramework.ts
@@ -8,7 +8,6 @@ import {
   IMidwayFramework,
   MidwayProcessTypeEnum,
   CommonFilterUnion,
-  CommonMiddleware,
   MiddlewareRespond,
   REQUEST_CTX_LOGGER_CACHE_KEY,
   ASYNC_CONTEXT_KEY,
@@ -344,27 +343,27 @@ export abstract class BaseFramework<
   ): Promise<void> {}
 
   public async applyMiddleware<R, N>(
-    lastMiddleware?: CommonMiddleware<CTX, R, N>
+    lastMiddleware?: CommonMiddlewareUnion<CTX, R, N>
   ): Promise<MiddlewareRespond<CTX, R, N>> {
-    if (!this.applicationContext.hasObject(ASYNC_CONTEXT_MANAGER_KEY)) {
-      const asyncContextManagerEnabled =
-        this.configService.getConfiguration('asyncContextManager.enable') ||
-        false;
-
-      const contextManager: AsyncContextManager = asyncContextManagerEnabled
-        ? this.bootstrapOptions?.asyncContextManager || new NoopContextManager()
-        : new NoopContextManager();
-
-      if (asyncContextManagerEnabled) {
-        contextManager.enable();
-      }
-      this.applicationContext.registerObject(
-        ASYNC_CONTEXT_MANAGER_KEY,
-        contextManager
-      );
-    }
-
     if (!this.composeMiddleware) {
+      if (!this.applicationContext.hasObject(ASYNC_CONTEXT_MANAGER_KEY)) {
+        const asyncContextManagerEnabled =
+          this.configService.getConfiguration('asyncContextManager.enable') ||
+          false;
+
+        const contextManager: AsyncContextManager = asyncContextManagerEnabled
+          ? this.bootstrapOptions?.asyncContextManager ||
+            new NoopContextManager()
+          : new NoopContextManager();
+
+        if (asyncContextManagerEnabled) {
+          contextManager.enable();
+        }
+        this.applicationContext.registerObject(
+          ASYNC_CONTEXT_MANAGER_KEY,
+          contextManager
+        );
+      }
       this.middlewareManager.insertFirst((async (ctx: any, next: any) => {
         // warp with context manager
         const rootContext = ASYNC_ROOT_CONTEXT.setValue(ASYNC_CONTEXT_KEY, ctx);
@@ -399,13 +398,15 @@ export abstract class BaseFramework<
       await this.filterManager.init(this.applicationContext);
     }
     if (lastMiddleware) {
+      lastMiddleware = Array.isArray(lastMiddleware)
+        ? lastMiddleware
+        : [lastMiddleware];
       return await this.middlewareService.compose(
-        [this.composeMiddleware, lastMiddleware],
+        [this.composeMiddleware, ...lastMiddleware],
         this.app
       );
-    } else {
-      return this.composeMiddleware;
     }
+    return this.composeMiddleware;
   }
 
   public getLogger(name?: string) {