+++ packages/core/src/common/dataSourceManager.ts
@@ -15,35 +15,38 @@ export abstract class DataSourceManager<T> {
 
   protected async initDataSource(options: any, appDir: string): Promise<void> {
     this.options = options;
-    if (options.dataSource) {
-      for (const dataSourceName in options.dataSource) {
-        const dataSourceOptions = options.dataSource[dataSourceName];
-        if (dataSourceOptions['entities']) {
-          const entities = new Set();
-          // loop entities and glob files to model
-          for (const entity of dataSourceOptions['entities']) {
-            if (typeof entity === 'string') {
-              // string will be glob file
-              const models = globModels(entity, appDir);
-              for (const model of models) {
-                entities.add(model);
-                this.modelMapping.set(model, dataSourceName);
-              }
-            } else {
-              // model will be add to array
-              entities.add(entity);
-              this.modelMapping.set(entity, dataSourceName);
-            }
-          }
-          dataSourceOptions['entities'] = Array.from(entities);
-        }
-        // create data source
-        await this.createInstance(dataSourceOptions, dataSourceName);
-      }
-    } else {
+    if (!options.dataSource) {
       throw new MidwayParameterError(
         'DataSourceManager must set options.dataSource.'
       );
+    }
+
+    for (const dataSourceName in options.dataSource) {
+      const dataSourceOptions = options.dataSource[dataSourceName];
+      if (dataSourceOptions['entities']) {
+        const entities = new Set();
+        // loop entities and glob files to model
+        for (const entity of dataSourceOptions['entities']) {
+          if (typeof entity === 'string') {
+            // string will be glob file
+            const models = globModels(entity, appDir);
+            for (const model of models) {
+              entities.add(model);
+              this.modelMapping.set(model, dataSourceName);
+            }
+          } else {
+            // model will be add to array
+            entities.add(entity);
+            this.modelMapping.set(entity, dataSourceName);
+          }
+        }
+        dataSourceOptions['entities'] = Array.from(entities);
+      }
+      // create data source
+      const opts: CreateInstanceOptions = {
+        cacheInstance: options.cacheInstance, // will default true
+      };
+      await this.createInstance(dataSourceOptions, dataSourceName, opts);
     }
   }
 
@@ -72,19 +75,27 @@ export abstract class DataSourceManager<T> {
    * @param dataSourceName
    */
   public async isConnected(dataSourceName: string): Promise<boolean> {
-    return this.checkConnected(this.getDataSource(dataSourceName));
+    const inst = this.getDataSource(dataSourceName);
+    return inst ? this.checkConnected(inst) : false;
   }
 
-  public async createInstance(config, clientName): Promise<T | void> {
+  public async createInstance(
+    config: any,
+    clientName: any,
+    options?: CreateInstanceOptions
+  ): Promise<T | void> {
+    const cache =
+      options && typeof options.cacheInstance === 'boolean'
+        ? options.cacheInstance
+        : true;
+
     // options.default will be merge in to options.clients[id]
-    config = extend(true, {}, this.options['default'], config);
-    const client = await this.createDataSource(config, clientName);
-    if (client) {
-      if (clientName) {
-        this.dataSource.set(clientName, client);
-      }
-      return client;
+    const configNow = extend(true, {}, this.options['default'], config);
+    const client = await this.createDataSource(configNow, clientName);
+    if (cache && clientName && client) {
+      this.dataSource.set(clientName, client);
     }
+    return client;
   }
 
   /**
@@ -136,4 +147,8 @@ export function globModels(globString: string, appDir: string) {
     }
   }
   return models;
+}
+
+export interface CreateInstanceOptions {
+  cacheInstance?: boolean | undefined;
 }