+++ packages/core/test/common/glob_dir_conflict/c.ts
@@ -1,0 +1,5 @@
+import { Provide } from '@midwayjs/decorator';
+
+@Provide()
+export class C {}
+