+++ packages/mikro/test/index.test.ts
@@ -4,7 +4,7 @@ import { close, createLightApp } from '@midwayjs/mock';
 import { IMidwayApplication } from '@midwayjs/core';
 
 describe('/test/index.test.ts', () => {
-  if (/v12/.test(process.version) || /v14/.test(process.version)) {
+  if (/v12/.test(process.version)) {
     it('should ignore mikro test', () => {
       console.log('not support current node version');
     });
@@ -12,7 +12,10 @@ describe('/test/index.test.ts', () => {
     it('should test base entity', async () => {
       cleanFile(join(__dirname, 'fixtures/base-fn-origin', 'test.sqlite'));
 
-      const app: IMidwayApplication = await createLightApp(join(__dirname, 'fixtures/base-fn-origin'), {});
+      const app: IMidwayApplication = await createLightApp(
+        join(__dirname, 'fixtures/base-fn-origin'),
+        {}
+      );
       const result = app.getAttr<string>('result');
 
       expect(result.includes('b1')).toBeTruthy();