+++ packages/web-koa/test/fixtures/base-app-query-parser/src/configuration.ts
@@ -1,0 +1,16 @@
+import { Configuration, App } from '@midwayjs/decorator';
+import { join } from 'path';
+
+@Configuration({
+  importConfigs: [
+    join(__dirname, './config'),
+  ]
+})
+export class ContainerConfiguration {
+
+  @App()
+  app;
+
+  async onReady() {
+  }
+}