+++ scripts/rollback/rollback-3.4.8.sh
@@ -1,0 +1,51 @@
+#!/bin/bash
+# timestamp: 1659445121727
+
+npm dist-tag add @midwayjs/fc-starter@3.4.7 latest
+npm dist-tag add @midwayjs/fc-starter@3.4.8 beta
+npm dist-tag add @midwayjs/code-dye@3.4.7 latest
+npm dist-tag add @midwayjs/code-dye@3.4.8 beta
+npm dist-tag add @midwayjs/cross-domain@3.4.7 latest
+npm dist-tag add @midwayjs/cross-domain@3.4.8 beta
+npm dist-tag add @midwayjs/faas@3.4.7 latest
+npm dist-tag add @midwayjs/faas@3.4.8 beta
+npm dist-tag add @midwayjs/http-proxy@3.4.7 latest
+npm dist-tag add @midwayjs/http-proxy@3.4.8 beta
+npm dist-tag add @midwayjs/security@3.4.7 latest
+npm dist-tag add @midwayjs/security@3.4.8 beta
+npm dist-tag add @midwayjs/static-file@3.4.7 latest
+npm dist-tag add @midwayjs/static-file@3.4.8 beta
+npm dist-tag add @midwayjs/upload@3.4.7 latest
+npm dist-tag add @midwayjs/upload@3.4.8 beta
+npm dist-tag add @midwayjs/version@3.4.7 latest
+npm dist-tag add @midwayjs/version@3.4.8 beta
+tnpm dist-tag add @midwayjs/fc-starter@3.4.7 latest
+tnpm dist-tag add @midwayjs/fc-starter@3.4.8 beta
+tnpm dist-tag add @midwayjs/code-dye@3.4.7 latest
+tnpm dist-tag add @midwayjs/code-dye@3.4.8 beta
+tnpm dist-tag add @midwayjs/cross-domain@3.4.7 latest
+tnpm dist-tag add @midwayjs/cross-domain@3.4.8 beta
+tnpm dist-tag add @midwayjs/faas@3.4.7 latest
+tnpm dist-tag add @midwayjs/faas@3.4.8 beta
+tnpm dist-tag add @midwayjs/http-proxy@3.4.7 latest
+tnpm dist-tag add @midwayjs/http-proxy@3.4.8 beta
+tnpm dist-tag add @midwayjs/security@3.4.7 latest
+tnpm dist-tag add @midwayjs/security@3.4.8 beta
+tnpm dist-tag add @midwayjs/static-file@3.4.7 latest
+tnpm dist-tag add @midwayjs/static-file@3.4.8 beta
+tnpm dist-tag add @midwayjs/upload@3.4.7 latest
+tnpm dist-tag add @midwayjs/upload@3.4.8 beta
+tnpm dist-tag add @midwayjs/version@3.4.7 latest
+tnpm dist-tag add @midwayjs/version@3.4.8 beta
+
+# Changes:
+
+#  - @midwayjs/fc-starter: 3.4.7 => 3.4.8
+#  - @midwayjs/code-dye: 3.4.7 => 3.4.8
+#  - @midwayjs/cross-domain: 3.4.7 => 3.4.8
+#  - @midwayjs/faas: 3.4.7 => 3.4.8
+#  - @midwayjs/http-proxy: 3.4.7 => 3.4.8
+#  - @midwayjs/security: 3.4.7 => 3.4.8
+#  - @midwayjs/static-file: 3.4.7 => 3.4.8
+#  - @midwayjs/upload: 3.4.7 => 3.4.8
+#  - @midwayjs/version: 3.4.7 => 3.4.8